requirements.txt
streamlit>=1.32.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pandas>=2.2.1
anthropic>=0.40.0
google-genai>=1.0.0
Pillow>=10.0.0
